+++ vite.config.js
@@ -2,6 +2,15 @@ import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 
 export default defineConfig({
+  plugins: [react()],
+  test: {
+    environment: "jsdom",
+    globals: true,
+    setupFiles: "./tests/setup.js",
+    env: {
+      IS_REACT_ACT_ENVIRONMENT: "true",
+    },
+  },
   build: {
     lib: {
       entry: "./src/lib/components/Modal.jsx",
@@ -18,5 +27,4 @@ export default defineConfig({
       },
     },
   },
-  plugins: [react()],
 });